+++ src/rule/index.ts
@@ -2,4 +2,4 @@ export {default as media}                    from "./MediaRule";
 export {default as platform}                 from "./PlatformRule";
 export {default as pseudo, createPseudoRule} from "./PseudoRule";
 
-export {and, or, createStyleRule} from "./StyleRule";
+export {and, or, not, createStyleRule} from "./StyleRule";