+++ src/rule/index.ts
@@ -1,5 +1,5 @@
-export {default as media}                    from "./MediaRule";
-export {default as platform}                 from "./PlatformRule";
-export {default as pseudo, createPseudoRule} from "./PseudoRule";
+export {default as media}                                            from "./MediaRule";
+export {default as platform}                                         from "./PlatformRule";
+export {default as pseudo, createPseudoRule, active, focus, disable} from "./PseudoRule";
 
 export {and, or, not, createStyleRule} from "./StyleRule";