+++ src/rule/index.ts
@@ -1,4 +1,5 @@
 export {default as media}                                             from "./MediaRule";
+export {default as rtl}                                               from "./RtlRule";
 export {default as platform}                                          from "./PlatformRule";
 export {default as pseudo, createPseudoRule, active, focus, disabled} from "./PseudoRule";