+++ src/rule/PseudoRule.ts
@@ -1,12 +1,12 @@
-import {createStyleRule, StyleRule} from "./StyleRule";
+import {createStyleRule, StyleRuleFunction} from "./StyleRule";
 
 const pseudo = createStyleRule<string>("pseudo", {
   check(options, session) {
-    return Boolean(session.context.pseudoClasses?.includes(options));
+    return Boolean(session?.pseudoClasses?.includes(options));
   },
 });
 
-export type PseudoRule = StyleRule & { type: string };
+export type PseudoRule = StyleRuleFunction & { type: string };
 
 export const createPseudoRule = (type: string): PseudoRule => Object.assign(() => {
   return pseudo(type);