+++ src/rule/index.ts
@@ -2,5 +2,6 @@ export {default as media}                                             from "./Me
 export {default as rtl}                                               from "./RtlRule";
 export {default as platform}                                          from "./PlatformRule";
 export {default as pseudo, createPseudoRule, active, focus, disabled} from "./PseudoRule";
+export {default as child}                                             from "./ChildRule";
 
 export {and, or, not, createStyleRule} from "./StyleRule";